Clarify the intention when getting or creating PropertyMetadata
Coverity did not like the possibility of the method returning null.
There was no risk for a NPE as the code was protected by the PropertyTypeCheck
Enum. However, it was overly complex, so splitting out the logic for
methods and fields reduces the risk of a regression in the future.

JAVA-2725