JAVA-2852: Log failure to get machine or process id at debug
The ObjectId class logs its failure to get either the machine identifer,
from NetworkInterface#getNetworkInterfaces, or the process
identifier, from JMX, at warning level, and includes the stack trace.

This is annoying for applications running on platforms where this occurs
(e.g. IBM JRE or Android's Dalvik), as it appears in the logs on every
application startup, there's no workaround, and the fallback behavior of
using SecureRandom is not a cause for concern.

This commit removes the stack trace from the log message and changes the
level from warn to debug