JAVA-2827: Re-enable maxAwaitTime test for AggregateOperation
The server only supports maxTimeMS on getMore for tailable cursors,
so added a $changeStream stage to the pipeline.

Note that this is not a particularly good test, since it doesn't
check to see whether maxTimeMS has been added to the getMore command.
It just assumes it has been and is just checking that no exception
is thrown.