Deprecate the mongodb-driver-async module
The mongodb-driver-async module API is based on callbacks, which are
difficult and error-prone to program against.  Since this API was
introduced, the industry has standardize around
http://www.reactive-streams.org/ and, more recently,
java.util.concurrent.Flow (introduced in Java 9).  Given this trend, it
is no longer advantageous to maintain the callback-based public API; it
is now deprecated and the intention is to remove it in the next
major release of the driver.

JAVA-3008