Remove connection string from log message
A connection string containing an unsupported option generates a log
message at warning level to the "org.mongodb.driver.uri" component. The
log message contains the full connection string.  As the connection
string may contain the credentials used to authenticate, it should not
be logged.  This commit removes the full connection string from the log
message, and instead just logs the name of the unsupported option.

JAVA-3093