Optimize check for an empty update document
The CRUD specification requires that an update contains at least one
field, in order to check for application bugs.  Because once the command
makes it to the server it can't be distinguished from a replace (where
an empty document is legal).

The current method of checking for an empty document is to call
BsonDocument#isEmpty.  But that has the unfortunate affect of hydrating
a BsonDocumentWrapper into a full BsonDocument.  We really want to avoid
that hit.  To do so, we need another way of checking for an empty
document.  We use the technique of wrapping the BsonWriter with a
FieldTracking BsonWriter, which tracks whether at least one field
is written.  BulkWriteBatch then checks that and throws if it's not.

This also fixes a bug in which that check was done for both update and
replace requests.  An empty replace document, while rare, is legal.

Thanks to @astral303 for finding this issue and for opening an initial
PR that provided inspiration for this fix.

JAVA-3328