JAVA-3201: Add workload executor (#560)
Add workload executor

The workload executor is an application which continuously runs a workload specification
until the process is killed.  The shutdown hook writes the rolled up results of all executed
operations to a results.json file in a specified directory.

This type of executor is useful to run while concurrently executing various cluster 
maintenance scenarios, to ensure that operations continue to execute successfully
during the maintenance window.