Remove assumption that jdk.net package is available
This commit removes any assumptions that jdk.net.ExtendedSocketOptions
is available at all at runtime, since the jdk.net package is not actually
part of the standard library (though in practice it is usually available).
If it's not available, no TCP keep-alive options will be set, but socket
creation will proceed instead of failing with a ClassNotFoundException.

JAVA-4005