JAVA-4034 Refactor read/write retries to accommodate for both `PoolClearedError` and CSOT (#782)
Changes in this commit do the following two major things:

1) These changes make selecting a server (`binding.get*ConnectionSource`)
and checking out a connection (`source.getConnection`) part of a retryable operation.
Such a change makes `MongoConnectionPoolClearedException`s part of the operation,
thus allowing us to retry the operation when the exception happens.

2) The retry limit (not more than two attempts) was hardcoded in the driver's code structure,
i.e., the code was written in such a way that it was describing the first attempt, then depending
on the outcome, the code for the second attempt was executed, with the code for the second attempt
mostly duplicating the code for the first attempt, but having different error handling. Such an approach
not only negatively affects the code readability, but also prevents changing the number of attempts,
let alone making decisions on whether to retry based not on an attempt limit.
The pseudocode in the specifications is also written this way, and at least some other drivers,
e.g., C#, Rust, took a similar approach to structure the code.

Changes in this commit represent the outcome of refactoring the code related to the retry logic.
In the new code the retry logic is decoupled from the logic of a retryable operation as much as possible.
Since our operations are quite complex and may decide to break retrying in the middle of an attempt,
e.g., because the server selected in the attempt does not support retries, the operation logic
is still aware of the fact that it may be retried. However, it is important to understand that
this awareness is a direct consequence of the business logic. It cannot be gotten rid of
regardless of the approach taken to structure the code. Operations that have simpler business logic
can be written in a retry-agnostic way without making changes to the retry framework
that was added as part of this commit.

With the changes in this commit, applying the client-side timeout (CSOT) to a retryable operation
is as simple as replacing the hard retry limit condition with a condition that checks
whether there is time left for attempting the operation again.

JAVA-4034