Support List<T> and Map<String, T> components in Java records
This enhances the support for any record component that is a
generic List or Map where the generic type of the List or Map
is a record or POJO.  It used to incorrectly decode to
org.bson.Document. Now it respects the generic type of the
List or Map when decoding.

JAVA-4667