Fix timing-related test failure in SessionsProseTest (#1325)
There are two tests in AbstractSessionsProseTest that rely on a running
mongocryptd process.  Each test starts and destroys the process, and it
appears that occasionally the destruction in the first test interferes with
the starting in the second test, and the second test fails to connect to
the mongocryptd process.

This commit attempts to work around this by starting a single mongocryptd
process and use it for both tests. Note that AbstractSessionsProseTest
has two subclasses and therefore the mongocryptd process is still actually
started and destroyed twice, it doesn't seem to be a problem in practice
because the two subclasses are in different modules and there is enough time
between their execution that there is interference.

JAVA-4998