Remove internal use of MongoNamespace.COMMAND_COLLECTION_NAME (#1782)
This is a remnant of the pre-OP_MSG wire protocol, which required
a synthetic collection name for most commands.  It's no longer needed,
except for the last remaining bit where the driver uses OP_QUERY for
the first message in the connection handshake.  

As part of this, RequestMessage and CommandMessage have been simplified.

JAVA-5939